feat: add an attribute-based instantiation of parallel execution nodes (#4503)
* feat: adds an attribute-based instantiation of parallel execution nodes

This PR expands the definition of an attribute on a physical plan node to a
collection of named simple interfaces:

    type PhysicalAttributes map[string]PhysicalAttr

The PhysicalAttr interface allows arbitrary data structures to be installed as
attribute values.

This PR then adds two attribute types:

 1. ParallelRun - indicates that a plan node should be instantiated multiple
    times when creating the execution graph. Each instantiation will handle
    some subset of the data. In the code we call this the parallel group. If
    the node is a source node, it should restrict source data to the specified
    group it has been assigned. If it is a non-source node, it should expect to
    receive parallel data (the group), and produce parallel data.

 2. ParallelMerge - indicates that plan node should merge the multiple
    instantiations of its predecessors. It receives parallel data and produces
    non-parallel data that can be sent to any existing transformation node.

For example, in the following graph, the RWA and TO nodes will carry the
ParallelRun attribute and will be instantiated multiple times. The merge node
will carry the ParallelMerge attribute and will be responsible for unioning the
subsets.

    read-window-aggregate -> to -> merge

A node has output attributes and required attributes. There is a new validation
step after planning that ensures predecessors output the attributes that it
requires. It is also possible for an attribute to require that all successors
require it. Using these tools we can validate that we always have a sound
convergence from parallel data to non-parallel data after the planner rewrites
the graph to create parallel execution graphs.

During instantiation of the execution graph from the plan, there are three
types of instantiations we need to support, multiplied by the two existing
cases -- source and non-source.

 1. Standard instantiation. These are non-parallel, non-merge nodes.
    There is a 1:1 relationship between planner node and execution graph
    node. Only a single copy of the node is made and it references the
    single copy of the predecessor nodes.

 2. Parallel instantiation. There are multiple copies of the node and of
    all predecessors. This applies to both source and non-source nodes.

 3. Merge instantiation. There is a single copy of the node, but multiple
    copies of the predecessors. These copies merge into the node.

* chore: formatting

* chore: staticcheck fix

* fix: update test case names and add comments in parallel_test.go

* feat: added a new kind of parallel-aware source node in executetest

Leave the original non-parallel source node as is and create a new one that
implements the necessary parallel-run behaviour. This eliminates the risk of
impacting existing known-good tests.

* chore: moved physical planner attributes to their own file

* fix: correct naming for parllel-merge kind

* fix: "parent" -> "predecessor" in parallel merge transformation

* fix: use flux.Error when validating attributes on the physical plan

* fix: use flux.Error for all errors generated in plan/physical.go

* fix: merge plan.ValidatePhysicalPlan and plan.ValidateAttributes

This change requires setting trigger specs in the parallel execution tests.

* chore: make DatasetIDFromNodeID non-public

* fix: use i+j in as the instance when computing parents dataset id

Using j is correct for the merge node. Using i is correct for the parallel-run
code. In this unified code that handles both cases we need to use i+j.
Note:
  * if i > 0 then j is always 0 and
  * if j > 0 then i is always 0

* chore: removed commented out trace message

* fix: create a specialized version of executetest.Table for parallel execution

Rather than modify executetest.Table, create a version
executetest.ParallelTable with a specialized Do function.