fix: add parallel merges with no successors to the results set
When a transformation has side effects and it terminates a pipeline there is no
yield added. This is the case with the to() transformation.

When we parallelize a to() transformation that terminates a pipeline, it no
longer terminates the pipeline. The merge node becomes the terminator.

Parallelization happens after yields are added, so we end up with a
non-side-effect node with no successors. The output is then lost because it
isn't added to the set of results.

To handle this case we also add a node to the result set when it is a merge
node with no successors.