refactor: move implicit result generation to executor
refs: #4465

Previously, terminal nodes (nodes without children) fell into 3
categories:

- those that _are literal yields_
- those that _have side-effects_
- everything else

In the planner, the first two cases are marked as "roots" which
contribute to the query's concurrecy quota.

The "everything else" case would add an implicit yield to the end of the
chain, which itself is marked as a "root." This essentially funnels the
"everything else" nodes into the "literal yields" category.

Later on in the executor, literal yields generate "results" and
separately, so are the nodes that _have side-effects_.

---

This diff aims to simplify the roles and responsibilities played by each
of the planner and executor.

At a high-level, the diff marks _all terminal nodes_ as "roots" in the
planner. In the executor, we handle the "everything else" nodes
similarly to how we handled the side-effecting nodes, by adding a
"result" for them to the execution state.

This seems to work fine for many simple cases but appears to break
things that depend on `TableObjectCompiler` which seems to no longer
produce results as expected. In tests, we'll see panics where we try to
consume the results from a "sub program" but get a `nil` back instead.

Other tests will need to be adjusted to account for the shape of the
resultant plan having fewer yields in it, or the name of the results
being slightly different.