feat: added a finish state to parallel-merge, always protect with lock
The parallel merge receives data from multiple input datasets. The execution
engine does not force these to be sequential, therefore we must protect
ourselves against calling finish on the single output dataset multiple times.
This PR adds a finished state tracker and uses the mutex to protect it in all
functions (Process included). In Process we can delay the locking to after the
output data is built. This doesn't appear to impact performance in practice.

This PR protects against recoverable panics when something in the execution
pipeline before the merge produces an error at execution time. It forces data
messages to immediately stop and guarantees a single output finish message.

Note that this transformation should really use the new narrow chunk processing
API. At this time, that is left for future work.