fix: spawn-containers was not launching anything for the "all" case
Looks like when the script was refactored to allow for individual
services to be launched, the "all" case was not well tested. As such,
when we didn't name any services to launch nothing would launch.

After fixing this issue, it seems the hdb container doesn't launch
properly in the current state of the script. Something about having
those docker commands in a function or the function being sent to the
background disconnects it from the main tty for the script causing it to
fail to perform the readiness check properly.

In practice it seemed this meant hdb would report "ready" too soon and
the seed data insert would fail, as well as the tests.

For now, I'm skipping hdb. We'll have to loop back to try and find a
good alternative approach.