fix: some containers take time to clean, ignore fails
Sometimes when we run `docker stop` the engine can take a beat before
the container is removed from the active list. The follow-up `docker rm`
can therefore try to destroy the container while it's stopping (and
`rm`'ing itself).

In such a case, ignore the failure. If the container truly sticks around
beyond these 2 commands, we'll see a different failure when we try to
relaunch the container with an identical name.