refactor: move implicit result generation to executor (#4545)
* refactor: move implicit result generation to executor

refs: #4465

Previously, terminal nodes (nodes without children) fell into 3
categories:

- those that _are literal yields_
- those that _have side-effects_
- everything else

In the planner, the first two cases are marked as "roots" which
contribute to the query's concurrecy quota.

The "everything else" case would add an implicit yield to the end of the
chain, which itself is marked as a "root." This essentially funnels the
"everything else" nodes into the "literal yields" category.

Later on in the executor, literal yields generate "results" and
separately, so are the nodes that _have side-effects_.

---

This diff aims to simplify the roles and responsibilities played by each
of the planner and executor.

At a high-level, the diff marks _all terminal nodes_ as "roots" in the
planner. In the executor, we handle the "everything else" nodes
similarly to how we handled the side-effecting nodes, by adding a
"result" for them to the execution state.

This seems to work fine for many simple cases but appears to break
things that depend on `TableObjectCompiler` which seems to no longer
produce results as expected. In tests, we'll see panics where we try to
consume the results from a "sub program" but get a `nil` back instead.

Other tests will need to be adjusted to account for the shape of the
resultant plan having fewer yields in it, or the name of the results
being slightly different.

* fix: ensure root nodes that are also leaf nodes produce results

A bug in the original refactor (moving away from implicit yields,
instead favoring implicit results) meant that execution graphs that
reduced to a single node would not have a result attached.

This diff fixes the problem continues the refactor by pulling the result
generating code into a method on the visitor so we can now call it from
the 3 spots where results need to be attached:

- yields
- leaves which are also roots
- roots

* test: update tests to no longer look for generated yields

With the move to skip adding yields during planning, instead letting
the executor add results directly, many tests needed to be adjusted to
no longer expect yields to be injected into the plan.

* refactor: executor gives errors when result names collide

To mimic the checks we do in the planner for yield name collisions, the
executor now checks the names of results as they are added to the
execution state.

* chore: rephrase/remove comments

* refactor: non-yields now generate results once at end of func

Given that (currently) all terminal nodes are instantiated only once, we
can simplify result generation a little more. Since we know the index
ahead of time (`0`) we can call `generateResult` right at the end before
the function returns.

* docs: comment on rationale for result names