chore: update lsp-types
While trying to migrate `flux-lsp` to use `tower-lsp`, there was an
issue where `tower-lsp` exported a newer version of `lsp_types` than is
used in the `lsp` feature of the flux crate. This meant that the build
failed because the `From`/`Into` traits aren't available in the
migration. This patch updates `lsp_types` to the same version.