feat(mqtt): enable dialer pool for everyone
Previously we had implemented a way to allow flux to reuse client
connections to mqtt across calls to `mqtt.publish`, but it was hidden
behind a feature flag pending verification of the new behavior.

After some manual testing, things appear to be in good shape. By
inspecting a mosquitto mqtt broker's server log, and toggling the flag
off/on, we could see that when the feature is enabled there's a single
client connection and serveral publishes (versus one connection per
publish when the flag is off).

This diff removes the feature flag, turning the new code path on for
everyone.