fix: skip parallel merge when selecting the result name based on side effects
If a node is a parallel merge then it was added by the planner and might be a
successor of a node with side effects. In order to preserve behaviour, skip
these parallel merge nodes and check if the predecessor has side effects. If
so, use that node name instead.