feat: add mqtt docker integration test (#4556)
* feat: add MQTT broker container for integ testing

* fix: spawn-containers was not launching anything for the "all" case

Looks like when the script was refactored to allow for individual
services to be launched, the "all" case was not well tested. As such,
when we didn't name any services to launch nothing would launch.

After fixing this issue, it seems the hdb container doesn't launch
properly in the current state of the script. Something about having
those docker commands in a function or the function being sent to the
background disconnects it from the main tty for the script causing it to
fail to perform the readiness check properly.

In practice it seemed this meant hdb would report "ready" too soon and
the seed data insert would fail, as well as the tests.

For now, I'm skipping hdb. We'll have to loop back to try and find a
good alternative approach.

* test: add test for mqtt.publish

This test currently fails because the client dependency isn't injected.

* chore: inject mqtt dialer for test dependencies

* fix: some containers take time to clean, ignore fails

Sometimes when we run `docker stop` the engine can take a beat before
the container is removed from the active list. The follow-up `docker rm`
can therefore try to destroy the container while it's stopping (and
`rm`'ing itself).

In such a case, ignore the failure. If the container truly sticks around
beyond these 2 commands, we'll see a different failure when we try to
relaunch the container with an identical name.

* chore: print message while NOT launching SAP HANA

* fix: launch SAP HANA without sending job to background

* chore: improve doc for how to shutdown all integ test containers

* chore: make generate

* fix: add span to context for all dependency tests (#4568)

Co-authored-by: Sean Brickley <sean@wabr.io>