chore(strings): refactor strings tests to use newer styles (#4570)
Prior to this change, both old and new style strings tests existed. The
old ones still existed because they were part of a benchmark. Those
benchmarks have been removed as we have another benchmark that covers
the same code paths and is stable without needing the stdlib to remain
unmodified.

Additionally the newer style tests were also updated to use the minimal
number of columns as these tests are not loaded into storage we do not
need the extra columns. This simplified the tests.