chore(checkfmt): fixes bug in checkfmt.sh for rust (#4572)
Prior to this change the checkfmt.sh script would find unformatted rust
files but would ignore the error. This was because of the scoping rules
used to make changing dirs easier. Now every check step is its own
function with clear scoping rules. A few rust files were not formatted
and those are fixed with this commit.