feat: take the concurrencyLimit from FF and keep it in the dependencies (#4564)
Pull the concurrencyLimit from feature flags at the start of the execution
process, before planning, and stash it in the ExecutionOptions, which live in
the ExecutionDependencies. From there it can then be modified by planner rules.
This allows parallelization rules to raise the limit if they parallelize a
query.

At the same time we move the defaultMemoryLimit from the planner and into the
execution options. We also move the computation of memory limit and concurrency
quota from the planner and into the executor.

Included are test cases covering the existing and the new method of
determining query concurrency.