refactor(memory): convert the allocator into an interface (#4539)
This converts the allocator to an interface so we can wrap it in
different locations. It expands on the arrow allocator to add a direct
access to the account method of the underlying resource allocator.

The original plan was to separate these two so the allocator would not
have an `Account()` method. This proved to be too difficult as the
`Account()` method is used in some very deep places in the code and
changing all of those locations was just not practical.