feat(cmd): merge and replace the main flux cli with the alternate one (#4611)
This merges the fmt and test commands into the flux cli that's been
getting tested in `internal/cmd` and then it promotes the one in
`internal/cmd` to `cmd`.

This removes the `execute` and `repl` commands and instead makes those
the default. When no parameters are used, the `repl` is started. When a
parameter is used, the script is executed. In addition, the script is
executed in a normal context rather than hacking it through the repl.
This should make the results of running a script from the CLI be more
similar to if the script were run in InfluxDB.