fix(execute): transformation transport adapter now always invokes finish
The transformationTransportAdapter would flush any buffers that were
cached and send them to be processed by the next transformation. If an
error occurred while this was happening, it would return an error
instead of invoking finish on the transformation. This caused some
transformations to not be finished when an error occurred and this would
lead to a deadlock.