feat: Accept string literals in the fields of a record type
For #4388 we need a way to describe a type (label) variable in place of a record field.
As type variables are described as a single, uppercase letter, mimicking that syntax will make it
impossible to describe the literal field name "A", "B", etc. By allowing a string literal in this location one can always use this syntax to describe these single letter field names and also describe any other fields which do not fit into an `identifier`, for example `"field with spaces"`, `"symbols#%^"`.