fix(stdlib): fix memory releases and add checked allocator to end to end tests
This commit adds the checked allocator to the end to end tests to ensure
that memory is actually released. It disables certain feature flags that
haven't been enabled yet that seem to have memory leak issues and
removes a package that has memory leak issues from the contrib package.

The table functions are a notable exception to these tests. The
interpreter was never updated to properly release values so the result
of the table functions does not ever get released when run outside of
the compiler. A special selection of tests have the memory tests
disabled to accomodate this.