perf(stdlib): improve aggregate window memory usage and polish the implementation (#4680)
This improves and polishes the builtin version of aggregate window.

First, instead of using indices, it reconstructs the time and numeric
column it is operating on. It turns out, for this case, it's more
efficient to reconstruct the times and values to be sorted. It removes
indirection and makes the overall algorithm more efficient.

Second, this polishes some parts like making sure it catches different
types in the same group key and removes some of the awkward interfaces.