fix: don't set BaseLocation.file to Some("") (#4684)
* fix: don't set BaseLocation.file to Some("")

A common pattern we have is this:

    parse_string("".to_string(), "from(...) |> range(...) |> filter(...)")

That first argument is supposed to serve as the filename and, in
general, we don't care about it. However, the `BaseLocation.file` has a
`#[serde(skip_serializing_if = "skip_string_option")]` on it, which
means that when we go over the serialization boundary,
`BaseLocation.file` is no longer `Some("")` but `None`. It's probably
not a huge deal, but when the result is inconsistent going over the
serialization boundary, that becomes an issue.

This patch fixes that by only assigning `BaseLocation.file` to a `Some`
value when the string is not empty. Otherwise, assign it as a `None`.

* test: add test for ast roundtrip