fix(universe/join): Treat JoinOpSpec.TableNames immutably
When a `join` is used in conjunction with a `tableFind`, the input tables
are copied and given new operation IDs. This causes problems in
`createMergeJoinTransformation` where we rely on a mapping of
`TableNames` to associate the inputs to `join` with the corresponding
datasets.

The assumption violated in this situation is we expect the number of
"parents" to be the same as the number of tracked "table names." When
this does not hold true, we panic. Formerly the panic came from an out
of bounds index, but with this diff this is more explicit.

Ideally, I think it would be better to perform a deep copy on the join's
`JoinOpSpec` for the sub-program we build via `tableFind` such that the
"sub join" is entirely different and can hold references to the "sub
inputs."
In practice this appears out of reach based on the information I saw
available in the code paths that matter. We'd need to add "join aware"
blocks to `TableObject.Operation()`, which seems to be responsible for
the "cloning" of the Operation.

The proposed fix here is to avoid altering the `TableNames` map for a
`JoinOpSpec` when this cloning happens. In effect, once a `JoinOpSpec`
is associated with a pair of input tables, this will be fixed. The
inputs may be cloned by `tableFind` but the cloned inputs  will no longer
be referred to as they are today.