fix: remove old entries in JoinOpSpec.TableNames before adding new
A previous attempt to fix a panic caused by the `TableNames` map
incorrectly growing when a `JoinOpSpec` is used repeatedly via
`tableFind` had "locked" the map after the initial entries had been
added.

This diff aims to solve the same problem by instead removing old entries
before adding new ones. This in effect should mean `TableNames` and
`params` should agree with each other.

While the truth is, we should not be reusing and mutating the OpSpec at
least now the fields that are changing with repeated use will agree with
each other.