refactor: remove `JoinOpSpec.TableNames`, favor `JoinOpSpec.params`
`JoinOpSpec.TableNames` mapped operationIDs to field names for the
`tables` argument to `join`.

Meanwhile `JoinOpSpec.params` held two separate fields:
- an array of strings (the field names of the object)
- an array of "operations" (the tables bound to those fields)

The `params` are in effect separate arrays of the keys/values seen in
the `TableNames` map, but holding references to the input table objects
rather than their operation IDs.

For the purposes of doing `join`'s work, the operation IDs are not
necessary, and in fact causes issues in situations like a when `join`
is used with `tableFind`.

Since the _extra information_ contained in the `TableNames` map is not
needed it's basically redundant.

This diff removes the field in favor of just referring to `params`
directly.