fix(universe/join): remove `JoinOpSpec.TableNames` and favor `JoinOpSpec.params` to stay consistent inside `tableFind` (#4698)
* fix(universe/join): Treat JoinOpSpec.TableNames immutably

When a `join` is used in conjunction with a `tableFind`, the input tables
are copied and given new operation IDs. This causes problems in
`createMergeJoinTransformation` where we rely on a mapping of
`TableNames` to associate the inputs to `join` with the corresponding
datasets.

The assumption violated in this situation is we expect the number of
"parents" to be the same as the number of tracked "table names." When
this does not hold true, we panic. Formerly the panic came from an out
of bounds index, but with this diff this is more explicit.

Ideally, I think it would be better to perform a deep copy on the join's
`JoinOpSpec` for the sub-program we build via `tableFind` such that the
"sub join" is entirely different and can hold references to the "sub
inputs."
In practice this appears out of reach based on the information I saw
available in the code paths that matter. We'd need to add "join aware"
blocks to `TableObject.Operation()`, which seems to be responsible for
the "cloning" of the Operation.

The proposed fix here is to avoid altering the `TableNames` map for a
`JoinOpSpec` when this cloning happens. In effect, once a `JoinOpSpec`
is associated with a pair of input tables, this will be fixed. The
inputs may be cloned by `tableFind` but the cloned inputs  will no longer
be referred to as they are today.

* chore: make generate

* test: add 2nd case with branching lookup

* chore: make generate

* test: 3rd case where the joins are defined inside the map

* chore: make generate

* fix: remove old entries in JoinOpSpec.TableNames before adding new

A previous attempt to fix a panic caused by the `TableNames` map
incorrectly growing when a `JoinOpSpec` is used repeatedly via
`tableFind` had "locked" the map after the initial entries had been
added.

This diff aims to solve the same problem by instead removing old entries
before adding new ones. This in effect should mean `TableNames` and
`params` should agree with each other.

While the truth is, we should not be reusing and mutating the OpSpec at
least now the fields that are changing with repeated use will agree with
each other.

* refactor: remove `JoinOpSpec.TableNames`, favor `JoinOpSpec.params`

`JoinOpSpec.TableNames` mapped operationIDs to field names for the
`tables` argument to `join`.

Meanwhile `JoinOpSpec.params` held two separate fields:
- an array of strings (the field names of the object)
- an array of "operations" (the tables bound to those fields)

The `params` are in effect separate arrays of the keys/values seen in
the `TableNames` map, but holding references to the input table objects
rather than their operation IDs.

For the purposes of doing `join`'s work, the operation IDs are not
necessary, and in fact causes issues in situations like a when `join`
is used with `tableFind`.

Since the _extra information_ contained in the `TableNames` map is not
needed it's basically redundant.

This diff removes the field in favor of just referring to `params`
directly.