test: add range to broken join test
During a recent refactor of join, some tests were updated to run as
proper acceptance tests.

This apparently broke in our nightly job, I think thanks to pushdowns
not present in vanilla flux.

This diff aims to fix the nightly suite by adding a `range` to the
`testing.load` call.