fix(stdlib): logic error in aggregate window rule and preserve whether table.fill is present
The aggregate window rule used the wrong condition for the infinite
window at the end because I was swapping the logical operators and
forgot one.

Additionally, `table.fill` needs to be included in the procedure spec.
This will be needed when this optimized aggregate window uses selectors,
but is mostly just needed to preserve the attribute so the influxdb push
downs can be properly formed no matter which direction the planner goes.