chore: Remove the experimental SetFinalizer allocator (#4737)
Every time we allocate an arrow array we check that we do not exceed the maximum memory limit.
Since this limit check is before the allocation step (and garbage collection) we inevitably
end up in a situation where the limit gets exceeded before the garbage collector has a chance
to run. Worse, go's garbage collector is highly incremental so even when it does run, we can
still end up allocating more than we collect, thereby exceeding the limit.

The only way I can see this allocator working is if the limit is so far above the amount of
memory actually being in use, which would let the garbage collector keep up. However in that case
the memory limit wouldn't really relate to how much memory were actually in use making it rather arbitrary.

Closes #4486