build: downgrade rust to v1.58 (#4753)
Fixes an issue with the OSS build.

Nb. the `Dockerfile_build` rust version (which normally has to be in
lock-step with the `rust-toolchain.toml` _because of reasons_) was
already at 1.58. This is why the version isn't changing in this diff.

Separately, could the drift between these 2 have contributed to the
break in the OSS build?