fix(stdlib/date): date.add and date.sub work correctly with timezones enabled
When timezones were enabled, date.add and date.sub did not correctly
work. To do time math, we convert from the local time to an equivalent
clock time in utc. We perform the math and then convert back to the
local time with the same clock time.

We did not correctly convert back to the local clock time when using
`date.add` and `date.sub`. This meant that they would convert to the utc
equivalent, perform the math, and then just never go back. This resulted
in incorrect times.

This was masked by the unit tests which only tested if a specific digit
of the time was correct. These test cases tested minutes which were not
affected by this bug. The tests have been updated to check the full time
instead of just a single digit.