fix: return a more helpful error message to the user when an http response exceeds maxResponseBody(100MB)
issue:
The flux http client limits response sizes to 100 MB. If a user runs a query that results in large http response, the resulting failure can manifest in a number of different error messages for the user, not all of which are helpful in effectively communicating the actual problem.
Other possible manifestation of this is truncated CSV, which produces a confusing error.

fixes: https://github.com/influxdata/flux/issues/4607