refactor: update parser to recognize `dynamic`
Added a test for `dynamic` as a parsable type, then minimal impls
for supporting code.

Still very unsure of the contracts I'm supposed to be satisfying along
this code path, but currently the dynamic type doesn't hold any
meaningful data, so there's a lot of "instantiate an empty struct"
happening along the way. If we end up needing to carry data along, these
will probably need to change.