feat: hacky, make json.parse able to handle heterogeneous arrays
This is not a good impl, but starts to head in the direction.

Updating the signature for `json.parse` to return a `dynamic` lets us
check the type system interaction statically. Specifically, we can look
at the static analysis applied when the output from parse is used or
passed around.

The (bad) impl itself simply shovels the marshaled values we get out of
the underlying parser into the array, which itself is typed using
`Dynamic` but this is only acting as a "straw man" type as enforcement
is now commented out!

A more correct impl would possibly wrap each value in a Dynamic type object.
Additionally, the root value itself (not just arrays) should probably be
wrapped. All the individual items that contribute to the return value,
perhaps? Not sure how deep or shallow we can afford to wrap.