feat: add a Dynamic values.Value
Dynamic value holds another value inside it, delegating to the inner for
basically everything.

Adding a new `Dynamic()` method to the `Value` interface for this means
implementing it for each of the original implementors. I've left these
as TODO panics since that's what IntelliJ spits out when prompted to
"implement remaining methods."