feat(execute): refactor the operator profile to be in the statistics
The operator profile is now always computed and it is part of the
`flux.Statistics` struct that is returned by the query. The operator
profile now uses that to return the profile instead of using Go channels
and the execution dependency injected into the context.

This allows the results of the operator profile to be present without it
being tied directly to the profiler return.

It has also been refactored to remove the use of channels for sending
span data and aggregating in a goroutine to instead just aggregate the
results as part of the source/transport code. This should simplify and
speed up those calculations as the overall time should be equivalent but
without the overhead of Go channels and only requiring a short-lived
lock for appending the source profiles in the executor.