feat: Replace unused implementations of Display for types
Ensures that we always go through the pretty.rs implementation.

(I checked flux-lsp and the only brekage is in a test due to the addition of whitespace).

BREAKING CHANGE

Some `std::fmt::Display` implementations have been removed in the `types` module. Convert to `MonoType` and use its `Display` implementation instead or write the formatting manually.