fix: ensure buffers are retained when copying an already buffered table
When the tables sent into `tableFind` include a yield, the result
generated by the yield can cause the tables to be released before
`tableFind` can use them.

`tableFind` will make its own copy, but prior to this change the copy
operation would hand back the input directly when it was already
buffered. In effect, this means we produce references to the same
underlying buffers without us incrementing the ref count.

With this diff, for the case where we try to copy an already buffered
table, we call its own `Copy()` method which will `Retain()` each of the
internal buffers.