fix(stdlib): vectorized map properly handles shadowed columns (#4819)
The vectorized version of map did not properly handle shadowed columns
the way the row based one worked. This copies over some of that code so
that it follows the correct logic. It also copies over the same logic
that is used to determine the column type when type inference does not
properly tell us the type.

This also fixes a bug in building objects that caused a bug in the
compiler. When a `with` is evaluated, the attributes from the original
record are copied over first and then overwritten values are assigned.
This overwriting did not properly release the values that were being
overwritten resulting in a memory leak.