fix: use atomic guards to ensure tableBuffer Do/Done are idempotent
Previously, `Do()` and `Done()` could run (both) resulting in the ref counts
for the underlying arrow arrays to be decremented twice instead of once per
table usage.

Normally this isn't an issue, but when a tableBuffer is copied for
reuse, the underlying buffers would be freed before they could be
properly copied.

Atomic guards are used to ensure `Done()` does nothing if called
repeatedly. Similarly, `Do()` will now return an error if the
tableBuffer is already used. This should prevent the "double free"
problem when `Done()` is deferred up front, while a `Do()` call is made
later on within a function body.