feat: Add testing.load on the converted tests
Input which used `testing.loadStorage` will use `testing.load` now in the same way that I have seen other tests use it. `testing.loadMem` which appear to be used on the expected data did not get converted as other converted tests do not appear to use `testing.load` on the expected data and there also does not appear to be any code that overrode the option.