fix: copy input table for benches using RowSelectorFuncBenchmarkHelper
Many benchmarks rely on a singleton table, `universe_test.NormalTable`.
The idea seemed to be that each iteration of a given bencher should make
a copy of this template table.

Previously, bugs in the tableBuffer implementation meant that a "used"
table would silently noop when you called `Do()` on it.

With fixes put in place to fix `Do()` such that it will error when called on
an already used table, problems in some select benchers were surfaced.
Specifically, benchers relying on `RowSelectorFuncBenchmarkHelper` would
"spend" the singleton table then other benchers would fail to run since
it's now an error to even `Copy()` tables that are already spent.

Prior to the fixes being put in place, I assume we'd have seen
considerably faster runs for iterations where the table was already used
since the work of the `Do()` call would effectively be skipped.

This change makes `RowSelectorFuncBenchmarkHelper` more similar to
`IndexSelectorFuncBenchmarkHelper` which correctly copied the input
table with each iteration of the bench.