fix: ensure buffers are retained when copying an already buffered table (#4820)
* fix: ensure buffers are retained when copying an already buffered table

When the tables sent into `tableFind` include a yield, the result
generated by the yield can cause the tables to be released before
`tableFind` can use them.

`tableFind` will make its own copy, but prior to this change the copy
operation would hand back the input directly when it was already
buffered. In effect, this means we produce references to the same
underlying buffers without us incrementing the ref count.

With this diff, for the case where we try to copy an already buffered
table, we call its own `Copy()` method which will `Retain()` each of the
internal buffers.

* test: add test to verify 2x `TableCopy` produces 2 copies

* fix: use atomic guards to ensure tableBuffer Do/Done are idempotent

Previously, `Do()` and `Done()` could run (both) resulting in the ref counts
for the underlying arrow arrays to be decremented twice instead of once per
table usage.

Normally this isn't an issue, but when a tableBuffer is copied for
reuse, the underlying buffers would be freed before they could be
properly copied.

Atomic guards are used to ensure `Done()` does nothing if called
repeatedly. Similarly, `Do()` will now return an error if the
tableBuffer is already used. This should prevent the "double free"
problem when `Done()` is deferred up front, while a `Do()` call is made
later on within a function body.

* test: refactor tests to account for the new Do/Done impls

* fix: copy input table for benches using RowSelectorFuncBenchmarkHelper

Many benchmarks rely on a singleton table, `universe_test.NormalTable`.
The idea seemed to be that each iteration of a given bencher should make
a copy of this template table.

Previously, bugs in the tableBuffer implementation meant that a "used"
table would silently noop when you called `Do()` on it.

With fixes put in place to fix `Do()` such that it will error when called on
an already used table, problems in some select benchers were surfaced.
Specifically, benchers relying on `RowSelectorFuncBenchmarkHelper` would
"spend" the singleton table then other benchers would fail to run since
it's now an error to even `Copy()` tables that are already spent.

Prior to the fixes being put in place, I assume we'd have seen
considerably faster runs for iterations where the table was already used
since the work of the `Do()` call would effectively be skipped.

This change makes `RowSelectorFuncBenchmarkHelper` more similar to
`IndexSelectorFuncBenchmarkHelper` which correctly copied the input
table with each iteration of the bench.

* chore: be more consistent

* fix: use atomic.LoadInt32 to read guard value

* chore: downgrade copy-after-use to a logging call instead of a panic

A concern with the prior impl was a table could be copied after it had
been drained, but wouldn't necessarily lead to a panic or error unless
it the dead copy _actually got used_.

In the future we could add something like the thing arrow has:

<https://github.com/apache/arrow/tree/e90472e35b40f58b17d408438bb8de1641bfe6ef/go/arrow/internal/debug>

This would allow us to have hard panics during test runs to help surface
correctness problems before they make their way out into production.

* feat: add a new `debug.Assert` based on Arrow's

Code added here is based on the v7 tag of Arrow:
<https://github.com/apache/arrow/tree/e90472e35b40f58b17d408438bb8de1641bfe6ef/go/arrow/internal/debug>

`debug.Assert` will noop if the `assert` build tag isn't set. Otherwise it
will panic with the specified message.

This mostly works like Rust's `debug_assert` in that we are currently
only setting the build tag during tests.

* chore: remove assert build tag from make test target (let CI set this)