feat(cmd/flux): check memory allocations to ensure no memory leaks in testcase
This changes the native flux test runner so it uses the checked
allocator to verify that no memory leaks occurred. This existed in the
old end to end tests, but was never ported over to testcase.

This only applies to the flux cli runner. It won't affect downstream
test runner implementations.