refactor: skip yields in input tables for cases like tableFind (#4869)
Table functions (ie tableFind) don't have any way to prioritize the use
of one result or another when they consume tables as input. For tables
that produce more than 1 result, the receiver would essentially use only
the first found result, which is non-deterministic.

In many cases, this can lead to errors at runtime where the output from
the table function does not match the expected schema. Perhaps worse,
it's possible you'd just get an incorrect answer rather than an error.

This refactor adds a new `skipYields` parameter to `spec.FromEvaluation`
which will produce a simplified spec when `true` by omitting any yields
encountered while walking the input tables.
With yields taken out of the picture, the only results we should see would
come from whatever the terminal node is, as well as from "side effecting"
functions such as the various flavors of `to` (influxdb, sql, kafka, etc).

In cases where the result count is > 1 thanks to side effecting
operations, in spite of stripping out yields, an error is returned while
`skipYields` is true.

The previous behavior of leaving all the yields in place should be used
for code paths that don't have the "single result" requirement.