test(plan): make planner test helper validate by default
Previously by default we would not perform validation steps
in physical planner tests since it presented a burden on the
test writer to make realistic plans and spend time on things
they are not actually testing.

This is an additive change that makes it so by default validation is
performed on the resulting plan, but it may be skipped if needed.
For tests that began failing due to validation failures, I have skipped
them.