fix: Add required fields to the signature of experimental/to
The documentation implies that these fields are mandatory to call this function and looking at the implementation it will indeed error if they are missing. Changing this is still technically a breaking change so maybe we want some caution before merging this? (https://github.com/influxdata/flux/pull/4773 could perhaps be used to allow an incremental rollout, or I could run this through the query log analyzer I made for https://github.com/influxdata/flux/pull/4776)