feat: add vectorRepeatValue; wrap literals with call expr, rewrite during eval
The caper is essentially to inject a fake function during the semantic
graph rewrite for vectorization, look for the function name while
evaluating call expressions, then branch to replace the node with a new
const vector evaluator (which holds a `vectorRepeatValue`).

The `vectorRepeatValue` is currently not very useful on its own. Access
to  the `Arr()` method will cause a panic, so callsites are updated to
type switch and avoid this. These callsites are where we need to branch
the behavior to account for the singleton value vs the array.