refactor(test): update test harness for error output (#4912)
The test harness has been updated so that more of the handling of the
results is handled by the test harness rather than the individual test
executor implementations.

This changes the interface so it returns a `flux.ResultIterator` rather
than handling the checking of the result.

This also removes the requirement for the error output yield to exist.
If the yield returns nothing, there's a chance it might not be emitted
at all because the csv encoder doesn't encode empty results. We might
want to fix that, but there's no way with the existing behavior and the
previous test harness didn't have this requirement.