fix(repl): replace extra boolean parameter for suggestions with repl options
This uses the option pattern to configure the repl with enabling
suggestions as the currently available option. This option defaults to
false. The `repl.New()` function has also been changed back to accept
only a single argument, but to also allow an optional set of options to
configure the REPL.