refactor(test): second refactor of the test executor (#4917)
The previous changes to the test executor worked in one environment, but
was too difficult to refactor with the defers in the second test
environment. This changes the flow back so the entire test is executed
inside of run and handles reading the result by using a callback
function instead.