feat(stdlib): deprecate experimental.csv
`experimental.csv` only provides a simple way to query a URL and parse
as CSV, now that `requests` package is complete its best to use and
compose this logic as needed. As it stands `experimental.csv` can't
check the HTTP status code and so its best this is left to the developer
writing the code.