refactor(skip): remove hardcoded skip list
The Flux test command no longer contains a hardcoded skip list.
Either the test was fixed such that it does not need to be skipped,
the test was deleted as it was no longer valuable, or a `skip` tag was
added thus skipping the test.

Additionally this change includes the testing.shouldError function taken
from here https://github.com/influxdata/flux/pull/4880