fix(stdlib): generate placeholder.go file when needed
Go's vendoring semantics mean we need all Flux packages to also be Go
packages. This change handles that logic. This is similar to the
flux_gen.go solution we had previously, but different in one key way.
The contents of placeholder.go are static. This means we do not need to
treat them as special binary files or otherwise. Additionally the new
name of placeholder.go is clear what its purpose is.