feat: allow physical plan attributes to contribute to formatter details (#4949)
In order to determine if a query plan has been parallelized we have been using
the presence of a partitionMerge node in the plan. With the emergence of a new
technique of merging in aggregateWindow nodes, and possibly soon pivot nodes as
well, we can no longer use this test.

This PR allows physical plan attributes to contribute to plan details. First
the spec is consulted, then any physical attributes that are present. Any node
with the parallel merge attribute can then indicate it is a merge node.