fix(stdlib): http writer skips invalid floats properly
The http writer now skips invalid floats properly and errors when the
field key is wrong with a message describing the problem.

Previously, we had the line protocol writer configured to ignore errors
when writing fields. But, when that happens, the line protocol writer
will return an error if all of the fields had an error with the very
opaque message "no serializable fields".

In order to avoid this, we needed to do some validation or control the
writing of line protocol a bit better. There is a line protocol v2
library that exposes the validation behavior so this change updates to
use that library instead.

The new and intended behavior is that if no fields can be written due to
NaN or Inf, we ignore the point entirely. If we cannot write it because
the field key is invalid, we error with a message specific to that. If
some fields can be written and some can't due to NaN or Inf, we only
ignore the fields that cannot be written.

This change also updates the influxdb provider to expose the
`influxdb.Metric` type with `influxdb.Tag` and `influxdb.Field`. These
were previously exposed through the v1 line protocol library. These are
presently exposed using a type alias to keep existing code compatible,
but existing code should switch to referencing these new types so we can
remove our current references to the v1 line protocol library.

The mqtt and kafka writers also use the v1 line protocol library and
have not been touched for this change. I would assume they have the same
problem as the existing `to()` behavior.

In addition to the above, this also modifies both influxdb `to()` and
`wideTo()` so that they skip NaN and Inf values and never send them to
the underlying writer. This should help prevent the need to handle these
in the dependency provider, but it's still useful to fix the dependency
provider to give better error messages anyway.