refactor: use specialized fns, fill array from repeat in map2
Abandons the -ArrLike interfaces to use specialized vector fns for each of
left or right consts.

Also fixes the case where no binary operation produces an array-backed
vector from a repeat. When a repeat "survives" to the point of `map2`
returning, use the value to build a full array of the same value.