refactor: constant folding now happens in binaryVectorEvaluator
By feeding the OperatorKind down through from the compiler into
binaryVectorEvaluator, we can check to see if the vectors we're
operating on are both constants (ie, `VectorRepeatValue`s).

In this case, we can short circuit the vectorized operator function and
use the regular row-based one to produce a new `VectorRepeatValue`.